requirements.txt
streamlit>=1.30.0
pandas
openpyxl
gspread
google-auth
plotly
